+++ regulation.hpp
@@ -8,8 +8,8 @@
 
 extern boolean chauffageAllume;
 extern boolean interrupteurPompe;
-extern double variationChauffage;
-extern double intervalCalculRegulation;
+extern long variationChauffage;
+extern long intervalCalculRegulation;
 
 extern void regulationInit(void);
 extern void regulationLoop(void);
